New: Sort avery list of education and experience by start-date in service layer, change type of education id to integer

+++ src/main/java/com/aras/cvcreator/service/EducationService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface EducationService {
     List<Education> findAll();
 
-    Education findById(Long id);
+    Education findById(Integer id);
 
     Education save(Education education);
